Extracted all limits out to their own file
Right now all limit attributes are defined as nested classes of
`Limit`. We've already done this for `Format` classes. Moving them out
to their own classes reduces bulk and makes it easier to find the class
you need.
