+++ CHANGELOG.md
@@ -5,6 +5,9 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
 
+### Changed
+- All limit classes are now stored in their own files in 'json-schema/attributes/limits'
+
 ### Fixed
 - Corrected the draft6 schema id to `http://json-schema.org/draft/schema#`
 - Rescue URI error when initializing a data string that contains a colon
