+++ lib/json-schema/attributes/limit.rb
@@ -48,132 +48,5 @@ module JSON
         raise NotImplementedError
       end
     end
-
-    class MinLengthAttribute < LimitAttribute
-      def self.acceptable_type
-        String
-      end
-
-      def self.limit_name
-        'minLength'
-      end
-
-      def self.error_message(schema)
-        "was not of a minimum string length of #{limit(schema)}"
-      end
-
-      def self.value(data)
-        data.length
-      end
-    end
-
-    class MaxLengthAttribute < MinLengthAttribute
-      def self.limit_name
-        'maxLength'
-      end
-
-      def self.error_message(schema)
-        "was not of a maximum string length of #{limit(schema)}"
-      end
-    end
-
-    class MinItemsAttribute < LimitAttribute
-      def self.acceptable_type
-        Array
-      end
-
-      def self.value(data)
-        data.length
-      end
-
-      def self.limit_name
-        'minItems'
-      end
-
-      def self.error_message(schema)
-        "did not contain a minimum number of items #{limit(schema)}"
-      end
-    end
-
-    class MaxItemsAttribute < MinItemsAttribute
-      def self.limit_name
-        'maxItems'
-      end
-
-      def self.error_message(schema)
-        "had more items than the allowed #{limit(schema)}"
-      end
-    end
-
-    class MinPropertiesAttribute < LimitAttribute
-      def self.acceptable_type
-        Hash
-      end
-
-      def self.value(data)
-        data.size
-      end
-
-      def self.limit_name
-        'minProperties'
-      end
-
-      def self.error_message(schema)
-        "did not contain a minimum number of properties #{limit(schema)}"
-      end
-    end
-
-    class MaxPropertiesAttribute < MinPropertiesAttribute
-      def self.limit_name
-        'maxProperties'
-      end
-
-      def self.error_message(schema)
-        "had more properties than the allowed #{limit(schema)}"
-      end
-    end
-
-    class NumericLimitAttribute < LimitAttribute
-      def self.acceptable_type
-        Numeric
-      end
-
-      def self.error_message(schema)
-        exclusivity = exclusive?(schema) ? 'exclusively' : 'inclusively'
-        format("did not have a %s value of %s, %s", limit_name, limit(schema), exclusivity)
-      end
-    end
-
-    class MaximumAttribute < NumericLimitAttribute
-      def self.limit_name
-        'maximum'
-      end
-
-      def self.exclusive?(schema)
-        schema['exclusiveMaximum']
-      end
-    end
-
-    class MaximumInclusiveAttribute < MaximumAttribute
-      def self.exclusive?(schema)
-        schema['maximumCanEqual'] == false
-      end
-    end
-
-    class MinimumAttribute < NumericLimitAttribute
-      def self.limit_name
-        'minimum'
-      end
-
-      def self.exclusive?(schema)
-        schema['exclusiveMinimum']
-      end
-    end
-
-    class MinimumInclusiveAttribute < MinimumAttribute
-      def self.exclusive?(schema)
-        schema['minimumCanEqual'] == false
-      end
-    end
   end
 end
