+++ lib/json-schema/attributes/limits/items.rb
@@ -1,0 +1,15 @@
+require 'json-schema/attributes/limit'
+
+module JSON
+  class Schema
+    class ItemsLimitAttribute < LimitAttribute
+      def self.acceptable_type
+        Array
+      end
+
+      def self.value(data)
+        data.length
+      end
+    end
+  end
+end
