+++ lib/json-schema/attributes/limits/minimum_inclusive.rb
@@ -1,0 +1,11 @@
+require 'json-schema/attributes/limits/minimum'
+
+module JSON
+  class Schema
+    class MinimumInclusiveAttribute < MinimumAttribute
+      def self.exclusive?(schema)
+        schema['minimumCanEqual'] == false
+      end
+    end
+  end
+end
