+++ lib/json-schema/attributes/limits/numeric.rb
@@ -1,0 +1,16 @@
+require 'json-schema/attributes/limit'
+
+module JSON
+  class Schema
+    class NumericLimitAttribute < LimitAttribute
+      def self.acceptable_type
+        Numeric
+      end
+
+      def self.error_message(schema)
+        exclusivity = exclusive?(schema) ? 'exclusively' : 'inclusively'
+        format("did not have a %s value of %s, %s", limit_name, limit(schema), exclusivity)
+      end
+    end
+  end
+end
